+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This application follows the [Semantic Versioning standard](https://semver.org/).
 
+## Version 0.1.1 (2022-xx-xx)
+
+### General
+- Implement `From<u16>` for `ApiErrorKind` and `From<ApiErrorKind>` for `u16`.
+- Implement `Display` for `ApiErrorKind`.
+
 ## Version 0.1.0 (2021-12-22)
 
 ### General